Fix bug by adding a float casting in RandomFloat

+++ TracerLib/PCG.cs
@@ -42,6 +42,6 @@ public class PCG
     /// <returns></returns>
     public float RandomFloat()
     {
-        return this.Random() / 0x100000000;
+        return this.Random() / (float)0x100000000;
     }
 }
